requirements.txt
autopep8==2.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.10.0
python-dotenv==0.21.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
WTForms==3.0.1
